Move debounce to its own file and add buffer sizing
This is just one of those things that probably only annoy me, but
anilist creates status messages when you watch episodes. If you spam the
episode plus button it would only create two request because of the
debounce. This would make a annoying situation if you spam clicked the
button and completed a series. You would get the following status messages
- watched episode x
- completed series x.

Now because of the buffer change we get three request that will make the
following status messages.
- watched episode x-maxEpisode - 1
- completed series x.