fix: 🐛 Don't cancel stats page requests
This bug effected every stat sub page like genres, tags etc. Here is how
the bug happened

1. Open up genres page
1.1 The genres page will now load old cached result and also send a new
request to get the fresh genre stats
1.2 The cache genre stats are passes down below and their id valued are
fetched from anilist
2. The fresh genre stats returned a result that will now replace the old
   genre stats and send a new update to sun component
2.1 The sub component will now cancel the first request for the media
id's and send the new one
2.2 The new one actually has the same id list as the old cached one
2.3 The api will always try to first look for matching cacheKey requests
and because we just send a request that is now cancelled, but the abort
signal has not yet registered, we will try to use that request
3. The request is cancel that this will cause the page to get null
   request

The fix was to not send a new fetch request if the list of ids are
identical. I also removed code that I don't think really did anything.